adding auth to all endpoints

+++ backend/routes/index.js
@@ -15,8 +15,9 @@ const router = express.Router();
 router.get("/", auth, async (req, res) => {
   try {
     const todos = await Todo.find({}).sort({ date: -1 });
+    const filteredTodos = todos.filter((todo) => todo.uid === req.user._id);
 
-    res.json(todos);
+    res.send(filteredTodos);
   } catch (error) {
     res.status(500).send(error.message);
     console.log(error.message);
@@ -36,7 +37,7 @@ router.get("/", auth, async (req, res) => {
 
 // ======================================================
 //POST
-router.post("/", async (req, res) => {
+router.post("/", auth, async (req, res) => {
   const schema = Joi.object({
     name: Joi.string().min(3).max(200).required(),
     author: Joi.string().min(3),
@@ -85,9 +86,13 @@ router.delete("/:id", auth, async (req, res) => {
     const todo = await Todo.findByIdAndDelete(req.params.id);
 
     if (!todo) return res.status(404).send("Todo not found...");
+
+    if (todo.uid !== req.user._id)
+      return res.status(401).send("Todo deletion failed. Not authorized...");
+
     const deletedTodo = await Todo.findByIdAndDelete(req.params.id);
 
-    res.send(todo);
+    res.send(deletedTodo);
   } catch (error) {
     res.send(500).send(error.message);
     console.log(error.message);
@@ -112,6 +117,9 @@ router.put("/:id", auth, async (req, res) => {
     const todo = await Todo.findById(req.params.id);
     if (!todo) return res.status(404).send("Todo not found...");
 
+    if (todo.uid !== req.user._id)
+      return res.status(401).send("Todo update failed. Not authorized...");
+
     const { name, author, isComplete, date } = req.body;
     const updatedTodo = await Todo.findByIdAndUpdate(
       req.params.id,
@@ -120,6 +128,7 @@ router.put("/:id", auth, async (req, res) => {
         author,
         isComplete,
         date,
+        uid,
       },
       { new: true }
     );
@@ -138,6 +147,11 @@ router.patch("/:id", auth, async (req, res) => {
     const todo = await Todo.findById(req.params.id);
 
     if (!todo) return res.status(404).send("Todo not found...");
+
+    if (todo.uid !== req.user._id)
+      return res
+        .status(401)
+        .send("Todo check/uncheck failed. Not authorized...");
 
     const updatedTodo = await Todo.findByIdAndUpdate(
       req.params.id,
